+++ internal/autodev/CHANGELOG.md
@@ -59,3 +59,6 @@
 
 ## [0.4.9+1781674543] - 2026-06-17
 - Test integration task
+
+## [0.4.9+1781679571] - 2026-06-17
+- Test integration task